Introduce 'threaduploadtexture'
It will use a dedicated opengl context on the pixmap reader thread
to upload textures there for images marked with
Image.asynchronous: true. This saves both the swizzling (for embedded)
and upload being done in the rendering thread.

It crashes on mac currently, so there is something still missing,
but I want the code in the repo to play more around with it.

Change-Id: I3579e2c2e0fb5705975dede19a8bab1480bf28ac
Reviewed-by: Samuel Rødal <samuel.rodal@digia.com>