Introduced 'mactexture'
Apple offers a couple of interesting texture options, namele client storage
and shared storage. Client Storage means that glTexImage2D does not need
to take a copy of the pixel data as we guarantee that the bits stay
alive for the lifetime of the texture. Shared Storage enables DMA transfer
of the pixels into graphics memory.

Change-Id: Ifd898d76510830130e9c9c0eba4823a5c49aee37
Reviewed-by: Samuel Rødal <samuel.rodal@digia.com>