Introduce 'nonpreservedtexture'
Provide a texture factory that does not keep a reference to the
QImage. This is intended to lower memory usage in the case of an
embedded system with only a single window that is never hidden or
re-exposed.

See also 8b50a9f00f6f5b693d18c59ff15968e2b902ebf8 in qtdeclarative.

Change-Id: If0a11f6bb37f4fb3b409a577609ebdfce73789b2
Reviewed-by: Gunnar Sletta <gunnar.sletta@digia.com>