Say hello to Qt.labs.shapes 0.1
The shapes plugin provies three element types.

1. TriangleSet - which encapsulates a set of triangles
including read/save logic to a very basic textual
format. Binary can of course be added later... It
also contains a set of path commands and uses Qt's
triangulator internally to convert a path into
triangles.

2. Polygon - Renders a flat shaded triangle set.

3. MaskedImage - Reads two triangle sets in addition
to an image and splits the image into a opaque and
blended part, making it possible to greatly reduce
the blending needed for drawing large bitmaps which are
largely opaque.

There is also a maskmaker tool in the tools subdirectory
which provides a simple means of generating masks for the
MaskedImage

Change-Id: Ia00f059fb81872e6ca1dca78baf3073c9581acab
Reviewed-by: Samuel Rødal <samuel.rodal@digia.com>