EGLImage/gralloc based textures
All the work relating to the texture upload is done on the
pixmap loader thread when the texture factory is created.
QSGTexture::bind() on the render thread then becomes a
zero-cost operation, which is very beneficial for large images
which tend to block the render thread for 10s of milliseconds.

Change-Id: I2d9eb15e64c435d58e454641dda7d588d6b236f4
Reviewed-by: Michael Brasser <michael.brasser@live.com>